Extract file list into reusable bit-file-list component
Split the attached-files list back out of bit-file-dropzone into a standalone
bit-file-list (name/size/delete + roving-focus focusDeleteAt), exported for
reuse outside the dropzone.
