+++ libs/components/src/file-upload/file-dropzone.component.html
@@ -63,30 +63,6 @@
     <!-- Hint sits between the drop area and the file list; the wrapper owns this layout. -->
     <ng-content select="bit-hint" />
 
-    @if (files().length > 0) {
-      <span [id]="listLabelId" class="tw-sr-only">{{ "uploadedFiles" | i18n }}</span>
-      <ul [attr.aria-labelledby]="listLabelId" class="tw-list-none tw-p-0 tw-m-0">
-        @for (file of files(); track $index) {
-          <li class="tw-flex tw-items-center tw-gap-3 tw-py-2 tw-border-b tw-border-border-light">
-            <bit-icon-tile icon="bwi-file" variant="subtle" size="sm" />
-            <div class="tw-flex tw-flex-col tw-flex-1 tw-min-w-0">
-              <bit-file-name [name]="file.name" class="tw-text-sm tw-font-semibold" />
-              <span class="tw-text-xs tw-text-muted">{{ formatFileSize(file.size) }}</span>
-            </div>
-            @if (!disabled()) {
-              <button
-                #deleteBtn
-                type="button"
-                bitIconButton="bwi-trash"
-                buttonType="dangerGhost"
-                size="small"
-                [label]="['delete' | i18n, file.name].join(' ')"
-                (click)="onRemove(file)"
-              ></button>
-            }
-          </li>
-        }
-      </ul>
-    }
+    <bit-file-list [files]="files()" [disabled]="disabled()" (fileRemoved)="onRemove($event)" />
   </div>
 </bit-form-field>
