+++ libs/components/src/file-upload/index.ts
@@ -1,2 +1,3 @@
 export { FileUploadComponent } from "./file-upload.component";
 export { FileDropzoneComponent } from "./file-dropzone.component";
+export { FileListComponent } from "./file-list.component";
